% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ab_si_conversions}
\alias{ab_si_conversions}
\title{Species site index conversion equations for Alberta and Saskatchewan}
\format{
A data frame with 12 rows (6 directed species pairs x 2 age bases):
\describe{
  \item{from}{source species code}
  \item{to}{target species code}
  \item{age_basis}{\code{"breast"} (Table 1, 50 years breast height age) or
    \code{"total"} (Table 1a, 50 years total age)}
  \item{intercept, slope}{RMA regression coefficients}
  \item{n_pairs}{number of matched species site index pairs used to fit}
  \item{n_plots}{total number of plots behind those pairs}
  \item{see}{standard error of estimate, metres}
  \item{pearson_r}{Pearson correlation coefficient}
  \item{p_value}{significance of the correlation}
}
}
\source{
Bjelanovic, I., and Comeau, P.G. 2019. Species SI conversion
  equations for Alberta and Saskatchewan. MGM Research Note #2019-2.
  University of Alberta, Edmonton, Alberta. Tables 1 and 1a.
}
\usage{
ab_si_conversions
}
\description{
Reduced major axis (RMA) regression coefficients for converting site index
between species at ecologically equivalent sites, from Bjelanovic and Comeau
(2019). Conversion is linear:
\code{SI_to = intercept + slope * SI_from}.
}
\seealso{
\code{\link{ab_si_to_si}}
}
\keyword{datasets}
