% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ab_si_ecosite}
\alias{ab_si_ecosite}
\title{Mean site index by natural subregion and ecosite}
\format{
A data frame with 118 rows:
\describe{
  \item{nsr}{natural subregion: \code{DM}, \code{CM}, \code{NM}, \code{LBH},
    \code{LF}, \code{UF}, \code{SA}, or \code{SASK} for Saskatchewan plots}
  \item{ecosite}{ecosite letter as used in the Alberta ecosite guides}
  \item{species}{\code{Aw}, \code{Sw}, \code{Pj} or \code{Pl}}
  \item{age_basis}{\code{"breast"} (Appendix 1) or \code{"total"}
    (Appendix 3)}
  \item{n}{number of plots}
  \item{si}{mean site index, metres}
  \item{sd}{standard deviation of site index, metres}
}
}
\source{
Bjelanovic, I., and Comeau, P.G. 2019. Estimating site index using
  ecosite and edatope in Alberta and Saskatchewan. MGM Research Note
  #2019-1. Appendices 1 and 3.
}
\usage{
ab_si_ecosite
}
\description{
Mean site index, standard deviation and sample size for each species in each
natural subregion and ecosite, from field and permanent sample plots in
Alberta and Saskatchewan. Only classes with three or more plots were
retained.
}
\seealso{
\code{\link{si_from_ecosite}}
}
\keyword{datasets}
